requirements.txt
Pillow==9.0.0
pygame==2.1.0
